# API.yaml
openapi: 3.0.3
info:
  title: Psy-Sup - OpenAPI 3.0
  version: '3.0'
tags:
  - name: user
    description: User information
  - name: post
    description: Post operations
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user
      operationId: createUser
      requestBody: 
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: successful operation
          headers:
            Location:
              description: The location of the newly updeted group
              schema:
                type: string
                format: uri
                example: '/user/{id}'
        '415':
          description: Unsupported Media Type
        '500':
          description: Server error
    get:
      tags:
        - user
      summary: List users
      description: ''
      operationId: getAllUsers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Server error
  /user/{id}:
    parameters:
    - name: id
      in: path
      description: ID of the user
      required: true
      schema:
        type: integer
    get:
      tags:
        - user
      summary: Get user info
      description: ''
      operationId: getUserById

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user
      operationId: updateUserById
      requestBody:
        description: Update an existent user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    delete:
      tags:
        - user
      summary: Delete post
      operationId: deleteUserByid
      parameters:
        - name: id
          in: path
          description: The user
          required: true
          schema:
            type: integer
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /post:
    post:
      tags:
        - post
      summary: Create post
      operationId: createPost
      requestBody: 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: successful operation
          headers:
            Location:
              description: The location of the newly updeted group
              schema:
                type: string
                format: uri
                example: '/user/{id}'
        '415':
          description: Unsupported Media Type
        '500':
          description: Server error
    get:
      tags:
        - post
      summary: List posts
      description: ''
      operationId: getAllPosts
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Server error
  /post/user/{id}:
    parameters:
    - name: id
      in: path
      description: ID of the user
      required: true
      schema:
        type: integer
    get:
      tags:
        - post
      summary: List posts of user
      description: ''
      operationId: getAllUserPosts
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Server error
  /post/{id}:
    parameters:
    - name: id
      in: path
      description: ID of the post
      required: true
      schema:
        type: integer
    get:
      tags:
        - post
      summary: Get post info
      description: ''
      operationId: getPostById

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'          
        '400':
          description: Invalid username supplied
        '404':
          description: Post not found
    put:
      tags:
        - post
      operationId: updatePostById
      summary: Update post
      requestBody:
        description: Update an existent post
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'          
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    delete:
      tags:
        - post
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deletePostById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          description: Invalid id supplied
        '404':
          description: Post not found
    
components:
  schemas:
    User:
      required:
        - username
        - email
        - phone
        - birthday
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        email:
          type: string
          format: email
          example: john@email.com
        phone:
          type: string
          example: '12345'
        birthday:
          type: string
          format: date
    Post:
      required:
        - userid
        - date
        - text
      properties:
        userid:
          type: integer
        text:
          type: string
        date:
          type: string
          format: date
          